+++ changelog.md
@@ -94,6 +94,7 @@ Codion Change Log
 - MenuBuilder, overloaded builder() methods removed.
 - MenuItemBuilder, overloaded builder() methods removed.
 - ToolBarBuilder, overloaded builder() methods removed.
+- LabelBuilder, overloaded builder() methods removed.
 ### is.codion.swing.framework.model
 - EntityComboBoxModel.Builder.EntityTypeBuilder and ConnectionProviderBuilder added.
 ### is.codion.swing.framework.ui